pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "anthropic>=0.84.0",
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "langchain-core>=1.2.18",
    "langgraph>=1.1.1",
    "lxml>=5.3.0",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "requests>=2.32.0",
    "sentence-transformers>=3.4.1",
    "yfinance>=1.2.0",
]
